+++ src/base/cs_time_control.c
@@ -220,11 +220,12 @@ cs_time_control_is_active(cs_time_control_t     *tc,
  *!
  * \brief Simple time control initialization based on time step options.
  *
- * \param[in]  tc        pointer to time control structure.
- * \param[in]  nt_start  start time step (or < 0 for unlimited)
- * \param[in]  nt_end    end time step (or < 0 for unlimited)
- * \param[in]  at_start  always active at start ?
- * \param[in]  at_start  always active at end ?
+ * \param[in]  tc           pointer to time control structure.
+ * \param[in]  nt_start     start time step (or < 0 for unlimited)
+ * \param[in]  nt_end       end time step (or < 0 for unlimited)
+ * \param[in]  nt_interval  time step interval (< 0 if no periodic output)
+ * \param[in]  at_start     always active at start ?
+ * \param[in]  at_start     always active at end ?
  */
 /*----------------------------------------------------------------------------*/
 
@@ -256,11 +257,12 @@ cs_time_control_init_by_time_step(cs_time_control_t  *tc,
  *!
  * \brief Simple time control initialization based on physical time options.
  *
- * \param[in]  tc        pointer to time control structure.
- * \param[in]  t_start   start time (or < 0 for unlimited)
- * \param[in]  t_end     end time (or < 0 for unlimited)
- * \param[in]  at_start  always active at start ?
- * \param[in]  at_start  always active at end ?
+ * \param[in]  tc          pointer to time control structure.
+ * \param[in]  t_start     start time (or < 0 for unlimited)
+ * \param[in]  t_end       end time (or < 0 for unlimited)
+ * \param[in]  t_interval  time interval (< 0 if no periodic output)
+ * \param[in]  at_start    always active at start ?
+ * \param[in]  at_start    always active at end ?
  */
 /*----------------------------------------------------------------------------*/